+++ example/lib/checkbox/checkboxes.dart
@@ -1,9 +1,9 @@
+import 'package:comanda_ui/comanda_ui.dart';
 import 'package:flutter/material.dart';
 
 import '../shared/component_card.dart';
 import '../shared/component_header.dart';
 import '../shared/component_scaffold.dart';
-import 'comanda_checkbox.dart';
 
 class CheckBoxes extends StatefulWidget {
   const CheckBoxes({Key? key}) : super(key: key);
@@ -27,48 +27,57 @@ class _CheckBoxesState extends State<CheckBoxes> {
         SizedBox(
           width: MediaQuery.of(context).size.width,
           child: Wrap(
+            spacing: 20,
             alignment: WrapAlignment.spaceBetween,
             children: [
               Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
                   ComponentHeader(title: leftCheckbox ? 'Ativo' : 'Inativo'),
-                  ComponentCard(
-                    child: Wrap(
-                      spacing: 20,
-                      children: [
-                        ComandaCheckBox(
-                          value: leftCheckbox,
-                          onChanged: (v) => changeLeftActive(v),
-                        ),
-                        ComandaCheckBox(
-                          text: 'Title',
-                          value: leftCheckbox,
-                          onChanged: (v) => changeLeftActive(v),
-                        ),
-                      ],
+                  SizedBox(
+                    width: MediaQuery.of(context).size.width <= 405 ? double.infinity : null,
+                    child: ComponentCard(
+                      child: Wrap(
+                        spacing: 20,
+                        runSpacing: 20,
+                        children: [
+                          ComandaCheckBox(
+                            value: leftCheckbox,
+                            onChanged: (v) => changeLeftActive(v),
+                          ),
+                          ComandaCheckBox(
+                            text: 'Title',
+                            value: leftCheckbox,
+                            onChanged: (v) => changeLeftActive(v),
+                          ),
+                        ],
+                      ),
                     ),
                   ),
                 ],
               ),
               Column(
-                crossAxisAlignment: CrossAxisAlignment.end,
+                crossAxisAlignment: MediaQuery.of(context).size.width <= 405 ? CrossAxisAlignment.start : CrossAxisAlignment.end,
                 children: [
                   ComponentHeader(title: rightCheckbox ? 'Ativo' : 'Inativo'),
-                  ComponentCard(
-                    child: Wrap(
-                      spacing: 20,
-                      children: [
-                        ComandaCheckBox(
-                          value: rightCheckbox,
-                          onChanged: (v) => changeRightActive(v),
-                        ),
-                        ComandaCheckBox(
-                          text: 'Title',
-                          value: rightCheckbox,
-                          onChanged: (v) => changeRightActive(v),
-                        ),
-                      ],
+                  SizedBox(
+                    width: MediaQuery.of(context).size.width <= 405 ? double.infinity : null,
+                    child: ComponentCard(
+                      child: Wrap(
+                        spacing: 20,
+                        runSpacing: 20,
+                        children: [
+                          ComandaCheckBox(
+                            value: rightCheckbox,
+                            onChanged: (v) => changeRightActive(v),
+                          ),
+                          ComandaCheckBox(
+                            text: 'Title',
+                            value: rightCheckbox,
+                            onChanged: (v) => changeRightActive(v),
+                          ),
+                        ],
+                      ),
                     ),
                   ),
                 ],