+++ example/lib/shared/component_appbar.dart
@@ -1,5 +1,6 @@
-import 'package:comanda_ui/shared/colors.dart';
-import 'package:comanda_ui/shared/weight.dart';
+
+
+import 'package:comanda_ui/comanda_ui.dart';
 import 'package:example/shared/utils/theme_preference.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';