+++ example/lib/shared/component_card.dart
@@ -1,4 +1,5 @@
-import 'package:comanda_ui/shared/colors.dart';
+
+import 'package:comanda_ui/comanda_ui.dart';
 import 'package:flutter/material.dart';
 
 class ComponentCard extends StatelessWidget {