+++ lib/components.dart
@@ -1,0 +1,7 @@
+library components;
+
+export 'src/components/comanda_button.dart';
+export 'src/components/comanda_checkbox.dart';
+export 'src/components/comanda_modal.dart';
+export 'src/components/comanda_snackbar.dart';
+export 'src/components/comanda_text_field.dart';