+++ lib/src/common/services/comanda_modal_service.dart
@@ -1,11 +1,5 @@
-import 'package:comanda_ui/shared/colors.dart';
-import 'package:comanda_ui/shared/weight.dart';
-import 'package:example/button/comanda_button.dart';
+import 'package:comanda_ui/comanda_ui.dart';
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-
-import '../../comanda.dart';
-import '../../modal/comanda_modal.dart';
 
 abstract class IComandaModalService {}