+++ lib/src/components/comanda_checkbox.dart
@@ -1,7 +1,8 @@
-import 'package:comanda_ui/shared/colors.dart';
-import 'package:comanda_ui/shared/weight.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+
+import '../common/theme/colors.dart';
+import '../common/theme/weight.dart';
 
 class ComandaCheckBox extends StatelessWidget {
   ComandaCheckBox({