+++ lib/src/components/comanda_snackbar.dart
@@ -1,6 +1,7 @@
-import 'package:comanda_ui/shared/weight.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+
+import '../common/theme/weight.dart';
 
 class ComandaSnackBar extends StatelessWidget {
   final String? title;