+++ lib/theme.dart
@@ -1,0 +1,5 @@
+library theme;
+
+export 'src/common/theme/theme.dart';
+export 'src/common/theme/weight.dart';
+export 'src/common/theme/colors.dart';